Barber/requirements.txt
aiogram>=3.6,<4.0
aiosqlite>=0.20.0
APScheduler>=3.10.4
python-dotenv>=1.0.1
